Move `get_fixture` fixture to root level
